dev-requirements.txt
-r requirements.txt
Faker==0.7.17
MarkupSafe==1.0
Werkzeug==0.12.2
asn1crypto==0.22.0
cffi==1.10.0
coverage==4.4.1
cryptography==1.9
decorator==4.0.11
django-debug-toolbar==1.8
entrypoints==0.2.3
factory-boy==2.8.1
ipdb==0.10.3
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
jedi==0.10.2
jinja2==2.9.6
jsonschema==2.6.0
jupyter-client==5.1.0
jupyter-core==4.3.0
mistune==0.7.4
nbconvert==5.2.1
nbformat==4.3.0
notebook==5.0.0
pandocfilters==1.4.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
ptyprocess==0.5.2
pyOpenSSL==17.1.0
pycparser==2.18
pygments==2.2.0
pyzmq==16.0.2
simplegeneric==0.8.1
sqlparse==0.2.3
terminado==0.6
testpath==0.3.1
tornado==4.5.1
traitlets==4.3.2
wcwidth==0.1.7
widgetsnbextension==2.0.0
